requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Deprecated==1.2.14
dimod==0.12.6
diskcache==5.6.1
dwave-cloud-client==0.10.6
dwave-drivers==0.4.4
dwave-greedy==0.3.0
dwave-hybrid==0.6.10
dwave-inspector==0.4.2
dwave-inspectorapp==0.3.1
dwave-neal==0.6.0
dwave-networkx==0.8.14
dwave-ocean-sdk==6.4.1
dwave-preprocessing==0.5.4
dwave-samplers==1.0.0
dwave-system==1.19.0
dwave-tabu==0.5.0
dwavebinarycsp==0.2.0
fasteners==0.18
Flask==2.3.2
homebase==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
minorminer==0.2.11
networkx==2.8.8
numpy==1.25.2
packaging==23.1
penaltymodel==1.0.2
plucky==0.4.3
pydantic==1.10.12
pyqubo==1.4.0
PySocks==1.7.1
python-dateutil==2.8.2
rectangle-packer==2.0.1
requests==2.31.0
scipy==1.11.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
